lexeme_entry disambig fixed
